+++ backend/clickhouse/errors.go
@@ -597,19 +597,36 @@ var ErrorObjectsTableConfig = model.TableConfig[modelInputs.ReservedErrorObjectK
 var ErrorsJoinedTableConfig = model.TableConfig[modelInputs.ReservedErrorsJoinedKey]{
 	TableName: "errors_joined_vw",
 	KeysToColumns: map[modelInputs.ReservedErrorsJoinedKey]string{
-		modelInputs.ReservedErrorsJoinedKeyBrowser:        "Browser",
-		modelInputs.ReservedErrorsJoinedKeyClientID:       "ClientID",
+		modelInputs.ReservedErrorsJoinedKeyBrowser:         "Browser",
+		modelInputs.ReservedErrorsJoinedKeyClientID:        "ClientID",
+		modelInputs.ReservedErrorsJoinedKeyEnvironment:     "Environment",
+		modelInputs.ReservedErrorsJoinedKeyEvent:           "Event",
+		modelInputs.ReservedErrorsJoinedKeyHasSession:      "HasSession",
+		modelInputs.ReservedErrorsJoinedKeyOsName:          "OSName",
+		modelInputs.ReservedErrorsJoinedKeySecureSessionID: "SecureSessionID",
+		modelInputs.ReservedErrorsJoinedKeyServiceName:     "ServiceName",
+		modelInputs.ReservedErrorsJoinedKeyServiceVersion:  "ServiceVersion",
+		modelInputs.ReservedErrorsJoinedKeyStatus:          "Status",
+		modelInputs.ReservedErrorsJoinedKeyTag:             "ErrorTagTitle",
+		modelInputs.ReservedErrorsJoinedKeyTimestamp:       "Timestamp",
+		modelInputs.ReservedErrorsJoinedKeyTraceID:         "TraceID",
+		modelInputs.ReservedErrorsJoinedKeyType:            "Type",
+		modelInputs.ReservedErrorsJoinedKeyVisitedURL:      "VisitedURL",
+	},
+	BodyColumn:   "Event",
+	ReservedKeys: modelInputs.AllReservedErrorsJoinedKey,
+}
+
+var BackendErrorObjectInputConfig = model.TableConfig[modelInputs.ReservedErrorsJoinedKey]{
+	KeysToColumns: map[modelInputs.ReservedErrorsJoinedKey]string{
 		modelInputs.ReservedErrorsJoinedKeyEnvironment:    "Environment",
 		modelInputs.ReservedErrorsJoinedKeyEvent:          "Event",
-		modelInputs.ReservedErrorsJoinedKeyHasSession:     "HasSession",
-		modelInputs.ReservedErrorsJoinedKeyOsName:         "OSName",
-		modelInputs.ReservedErrorsJoinedKeyServiceName:    "ServiceName",
-		modelInputs.ReservedErrorsJoinedKeyServiceVersion: "ServiceVersion",
-		modelInputs.ReservedErrorsJoinedKeyTag:            "ErrorTagTitle",
-		modelInputs.ReservedErrorsJoinedKeyType:           "Type",
-		modelInputs.ReservedErrorsJoinedKeyVisitedURL:     "VisitedURL",
+		modelInputs.ReservedErrorsJoinedKeyHasSession:     "SessionSecureID",
+		modelInputs.ReservedErrorsJoinedKeyServiceName:    "Service.Name",
+		modelInputs.ReservedErrorsJoinedKeyServiceVersion: "Service.Version",
 		modelInputs.ReservedErrorsJoinedKeyTimestamp:      "Timestamp",
-		modelInputs.ReservedErrorsJoinedKeyStatus:         "Status",
+		modelInputs.ReservedErrorsJoinedKeyType:           "Type",
+		modelInputs.ReservedErrorsJoinedKeyVisitedURL:     "URL",
 	},
 	BodyColumn:   "Event",
 	ReservedKeys: modelInputs.AllReservedErrorsJoinedKey,
@@ -623,7 +640,7 @@ var errorsSampleableTableConfig = sampleableTableConfig[modelInputs.ReservedErro
 }
 
 func ErrorMatchesQuery(errorObject *model2.BackendErrorObjectInput, filters listener.Filters) bool {
-	return matchesQuery(errorObject, ErrorsJoinedTableConfig, filters)
+	return matchesQuery(errorObject, BackendErrorObjectInputConfig, filters, listener.OperatorAnd)
 }
 
 func (client *Client) ReadErrorsMetrics(ctx context.Context, projectID int, params modelInputs.QueryInput, column string, metricTypes []modelInputs.MetricAggregator, groupBy []string, nBuckets *int, bucketBy string, limit *int, limitAggregator *modelInputs.MetricAggregator, limitColumn *string) (*modelInputs.MetricsBuckets, error) {
@@ -640,14 +657,14 @@ func (client *Client) QueryErrorObjectsHistogram(ctx context.Context, projectId 
 		return nil, nil, err
 	}
 
-	selectCols := fmt.Sprintf("%s(Timestamp, '%s') as time, count() as count", aggFn, location.String())
-
-	orderBy := fmt.Sprintf("1 WITH FILL FROM %s(?, '%s') TO %s(?, '%s') STEP 1", aggFn, location.String(), aggFn, location.String())
-
-	sb, err := readErrorsObjects(selectCols, params, projectId, "1", orderBy)
+	sb, err := readErrorsObjects(params, projectId)
 	if err != nil {
 		return nil, nil, err
 	}
+
+	sb.Select(fmt.Sprintf("%s(Timestamp, '%s') as time, count() as count", aggFn, location.String()))
+	sb.GroupBy("1")
+	sb.OrderBy(fmt.Sprintf("1 WITH FILL FROM %s(?, '%s') TO %s(?, '%s') STEP 1", aggFn, location.String(), aggFn, location.String()))
 
 	sql, args := sb.BuildWithFlavor(sqlbuilder.ClickHouse)
 	args = append(args, *options.Bounds.StartDate, *options.Bounds.EndDate)
@@ -687,6 +704,8 @@ func (client *Client) QueryErrorGroups(ctx context.Context, projectId int, count
 		return nil, 0, err
 	}
 
+	sb.Select("ID, count() OVER() AS total")
+	sb.OrderBy("UpdatedAt DESC, ID DESC")
 	sb.Limit(count)
 	sb.Offset(offset)
 
@@ -712,12 +731,11 @@ func (client *Client) QueryErrorGroups(ctx context.Context, projectId int, count
 
 func readErrorGroups(params modelInputs.QueryInput, projectId int) (*sqlbuilder.SelectBuilder, error) {
 	sb := sqlbuilder.NewSelectBuilder()
-	sb.From(ErrorGroupsTableConfig.TableName)
-	sb.Select("ID, count() OVER() AS total")
+	sb.From(fmt.Sprintf("%s FINAL", ErrorGroupsTableConfig.TableName))
 
 	sbInner := sqlbuilder.NewSelectBuilder()
 	sbInner.Select("ErrorGroupID")
-	sbInner.From(ErrorsJoinedTableConfig.TableName)
+	sbInner.From(fmt.Sprintf("%s FINAL", ErrorsJoinedTableConfig.TableName))
 	sbInner.Where(sbInner.Equal("ProjectId", projectId))
 
 	sbInner.Where(sbInner.LessEqualThan("Timestamp", params.DateRange.EndDate)).
@@ -730,16 +748,15 @@ func readErrorGroups(params modelInputs.QueryInput, projectId int) (*sqlbuilder.
 	return sb, nil
 }
 
-func readErrorsObjects(selectCols string, params modelInputs.QueryInput, projectId int, groupBy string, orderBy string) (*sqlbuilder.SelectBuilder, error) {
+func readErrorsObjects(params modelInputs.QueryInput, projectId int) (*sqlbuilder.SelectBuilder, error) {
 	sb := sqlbuilder.NewSelectBuilder()
-	sb.Select(selectCols)
 	sb.From(fmt.Sprintf("%s FINAL", ErrorsJoinedTableConfig.TableName))
 	sb.Where(sb.Equal("ProjectId", projectId))
 
-	parser.AssignSearchFilters(sb, params.Query, ErrorsJoinedTableConfig)
+	sb.Where(sb.LessEqualThan("Timestamp", params.DateRange.EndDate)).
+		Where(sb.GreaterEqualThan("Timestamp", params.DateRange.StartDate))
 
-	sb.OrderBy(orderBy)
-	sb.GroupBy(groupBy)
+	parser.AssignSearchFilters(sb, params.Query, ErrorsJoinedTableConfig)
 
 	return sb, nil
 }