+++ frontend/src/pages/Graphing/components/BarChart.tsx
@@ -3,19 +3,21 @@ import {
 	BarChart as RechartsBarChart,
 	CartesianGrid,
 	ResponsiveContainer,
+	Tooltip,
 	XAxis,
 	YAxis,
 } from 'recharts'
 
 import {
-	ChartProps,
 	CustomXAxisTick,
 	CustomYAxisTick,
-	getFormatter,
+	getColor,
+	getCustomTooltip,
+	getTickFormatter,
 	GROUP_KEY,
+	InnerChartProps,
 	isActive,
 	SeriesInfo,
-	strokeColors,
 } from '@/pages/Graphing/components/Graph'
 
 export type BarDisplay = 'Grouped' | 'Stacked'
@@ -51,13 +53,13 @@ export const BarChart = ({
 	series,
 	spotlight,
 	viewConfig,
-}: ChartProps<BarChartConfig> & SeriesInfo) => {
-	const xAxisTickFormatter = getFormatter(xAxisMetric, data?.length)
-	const yAxisTickFormatter = getFormatter(yAxisMetric)
+}: InnerChartProps<BarChartConfig> & SeriesInfo) => {
+	const xAxisTickFormatter = getTickFormatter(xAxisMetric, data?.length)
+	const yAxisTickFormatter = getTickFormatter(yAxisMetric)
 
 	return (
 		<ResponsiveContainer>
-			<RechartsBarChart data={data} barCategoryGap={0}>
+			<RechartsBarChart data={data} barCategoryGap={1}>
 				<XAxis
 					dataKey={xAxisMetric}
 					fontSize={10}
@@ -75,6 +77,12 @@ export const BarChart = ({
 					height={12}
 					type={xAxisMetric === GROUP_KEY ? 'category' : 'number'}
 					domain={['auto', 'auto']}
+				/>
+
+				<Tooltip
+					content={getCustomTooltip(xAxisMetric, yAxisMetric)}
+					cursor={{ fill: '#C8C7CB', fillOpacity: 0.5 }}
+					isAnimationActive={false}
 				/>
 
 				<YAxis
@@ -110,7 +118,7 @@ export const BarChart = ({
 							<Bar
 								key={key}
 								dataKey={key}
-								fill={strokeColors[idx % strokeColors.length]}
+								fill={getColor(idx)}
 								maxBarSize={30}
 								isAnimationActive={false}
 								stackId={