+++ frontend/src/pages/Traces/utils.ts
@@ -60,8 +60,10 @@ export const getTraceDurationString = (duration: number) => {
 		return `${secondString} ${millisecondString}`
 	} else if (millisecondString) {
 		return millisecondString
-	} else {
+	} else if (microsecondString) {
 		return microsecondString
+	} else {
+		return `${nanoseconds}ns`
 	}
 }