+++ frontend/src/routers/ProjectRouter/ApplicationRouter.tsx
@@ -2,7 +2,6 @@ import { useAuthContext } from '@authentication/AuthContext'
 import AlertsRouter from '@pages/Alerts/AlertsRouter'
 import LogAlertsRouter from '@pages/Alerts/LogAlert/LogAlertRouter'
 import { CanvasPage } from '@pages/Buttons/CanvasV2'
-import DashboardsRouter from '@pages/Dashboards/DashboardsRouter'
 import { useErrorSearchContext } from '@pages/Errors/ErrorSearchContext/ErrorSearchContext'
 import ErrorsV2 from '@pages/ErrorsV2/ErrorsV2'
 import IntegrationsPage from '@pages/IntegrationsPage/IntegrationsPage'
@@ -12,19 +11,21 @@ import { useSearchContext } from '@pages/Sessions/SearchContext/SearchContext'
 import { SetupRouter } from '@pages/Setup/SetupRouter/SetupRouter'
 import { usePreloadErrors, usePreloadSessions } from '@util/preload'
 import React, { Suspense } from 'react'
-import { Route, Routes } from 'react-router-dom'
+import { Navigate, Route, Routes } from 'react-router-dom'
 
 import { DEMO_PROJECT_ID } from '@/components/DemoWorkspaceButton/DemoWorkspaceButton'
 import { RelatedResourcePanel } from '@/components/RelatedResources/RelatedResourcePanel'
 import { useNumericProjectId } from '@/hooks/useProjectId'
 import { SignInRedirect } from '@/pages/Auth/SignInRedirect'
-import { GraphingEditor } from '@/pages/Graphing/GraphingEditor'
+import DashboardRouter from '@/pages/Graphing/DashboardRouter'
 import { SettingsRouter } from '@/pages/SettingsRouter/SettingsRouter'
 import { TracePanel } from '@/pages/Traces/TracePanel'
 import { TracesPage } from '@/pages/Traces/TracesPage'
 
 const Buttons = React.lazy(() => import('../../pages/Buttons/Buttons'))
 const HitTargets = React.lazy(() => import('../../pages/Buttons/HitTargets'))
+
+const BASE_PATH = 'sessions'
 
 const ApplicationRouter: React.FC = () => {
 	const { projectId } = useNumericProjectId()
@@ -101,16 +102,14 @@ const ApplicationRouter: React.FC = () => {
 						/>
 						{isHighlightAdmin && (
 							<Route
-								path="metrics/*"
-								element={
-									<Suspense fallback={null}>
-										<GraphingEditor />
-									</Suspense>
-								}
+								path="dashboards/*"
+								element={<DashboardRouter />}
 							/>
 						)}
-
-						<Route path="*" element={<DashboardsRouter />} />
+						<Route
+							path="*"
+							element={<Navigate to={BASE_PATH} replace />}
+						/>
 					</>
 				) : (
 					<Route path="*" element={<SignInRedirect />} />